+++ tests/testthat/test-fetch-survey.R
@@ -1,17 +1,77 @@
-context("Download a survey from qualtRics and pull it into R using the fetch_survey() function") # nolint
+context("Download a survey from qualtRics and pull it into R using the fetch_survey() function")
 
-test_that("fetch_survey() throws error", {
-  qualtrics_api_credentials(api_key = "1234",
-                            base_url = "https://yourdatacenterid.qualtrics.com")
-  # Query fake ID with generic root url
-  expect_error(
-    qualtRics::fetch_survey("1234"),
-    "you may not have the\nrequired authorization"
-  )
+test_that("fetch_survey() returns survey with default params", {
+
+  skip_on_cran()
+
+  qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
+
+  vcr::use_cassette("fetch_survey", {
+    x <- fetch_survey("SV_3gbwq8aJgqPwQDP")
+  })
+
+  expect_s3_class(x, c("spec_tbl_df", "tbl_df","tbl","data.frame"))
+  expect_named(x, c("StartDate", "EndDate", "Status", "IPAddress", "Progress",
+                    "Duration (in seconds)", "Finished", "RecordedDate",
+                    "ResponseId", "RecipientLastName", "RecipientFirstName",
+                    "RecipientEmail", "ExternalReference", "LocationLatitude",
+                    "LocationLongitude", "DistributionChannel", "UserLanguage",
+                    "Q63", "Q16", "Q17", "Q18", "Q19", "Q22", "SolutionRevision"))
+  expect_type(x$StartDate, "double")
+  expect_type(x$Status, "character")
+  expect_type(x$`Duration (in seconds)`, "double")
+  expect_type(x$Finished, "logical")
+  expect_type(x$ResponseId, "character")
+  expect_type(x$Q63, "integer")
+  expect_type(x$Q22, "character")
+
 })
 
-test_that("using fetch_survey() with a base URL that doesn't end with '.qualtrics.com' fails", { # nolint
-  # Store dummy key
+test_that("fetch_survey() returns survey with custom params", {
+
+  skip_on_cran()
+
+  qualtrics_api_credentials(api_key = "1234", base_url = "t.qualtrics.com")
+
+  vcr::use_cassette("fetch_survey_custom", {
+    x <- fetch_survey(
+      "SV_56icaa9YAafpAqx",
+      force_request = TRUE,
+      start_date = "2020-01-01",
+      end_date = "2020-03-01",
+      unanswer_recode = 999,
+      limit = 15,
+      include_questions = c("QID9", "QID21"),
+      breakout_sets = FALSE
+    )
+  })
+
+  expect_s3_class(x, c("spec_tbl_df", "tbl_df","tbl","data.frame"))
+  expect_equal(nrow(x), 15)
+  expect_named(x, c("StartDate", "EndDate", "Status",
+                    "Progress", "Duration (in seconds)", "Finished",
+                    "RecordedDate", "ResponseId", "DistributionChannel",
+                    "UserLanguage", "Q3.2", "Q4.1",
+                    "SolutionRevision",
+                    "Q3.8 - Parent Topics",
+                    "Q3.8 - Sentiment Polarity",
+                    "Q3.8 - Sentiment Score",
+                    "Q3.8 - Sentiment",
+                    "Q3.8 - Topic Sentiment Label",
+                    "Q3.8 - Topic Sentiment Score",
+                    "Q3.8 - Topics"))
+  expect_type(x$StartDate, "double")
+  expect_type(x$Status, "character")
+  expect_type(x$`Duration (in seconds)`, "double")
+  expect_type(x$Finished, "logical")
+  expect_type(x$ResponseId, "character")
+  expect_type(x$`Q3.2`, "integer")
+  expect_type(x$`Q4.1`, "integer")
+
+})
+
+test_that("using fetch_survey() with a base URL that doesn't end with '.qualtrics.com' fails", {
+
   qualtrics_api_credentials(api_key = "1234", base_url = "abcd")
   # This should error on check_params()
   expect_error(
@@ -20,11 +80,11 @@ test_that("using fetch_survey() with a base URL that doesn't end with '.qualtric
   ) # nolint
 })
 
-test_that("fetch_survey() reads a stored survey in temporary directory if exists", { # nolint
+test_that("fetch_survey() reads a stored survey in temporary directory if exists", {
   # Register dummy key
   qualtrics_api_credentials(
     api_key = "1234",
-    base_url = "https://yourdatacenterid.qualtrics.com"
+    base_url = "yourdatacenterid.qualtrics.com"
   )
   # Store RDS file
   data <- "SUCCESS"
@@ -37,12 +97,13 @@ test_that("fetch_survey() reads a stored survey in temporary directory if exists
     qualtRics::fetch_survey("surveyID"),
     "Found an earlier download for survey with id surveyID. Loading this file.
 Set 'force_request' to TRUE if you want to override this"
-  ) # nolint
+  )
 })
 
 test_that("Save directory exists for fetch_survey()", {
   expect_error(
-    qualtRics::fetch_survey("1234",
+    qualtRics::fetch_survey(
+      "1234",
       save_dir = "/users/jasper/desktop/idonotexist"
     ),
     "does not exist."
@@ -66,7 +127,7 @@ test_that("Handle convert and label conditions", {
 test_that("unanswer_recode is integer-ish", {
   qualtrics_api_credentials(
     api_key = "1234",
-    base_url = "yourdatacenterid.qualtrics.com"
+    base_url = "https://yourdatacenterid.qualtrics.com"
   )
   # Call fetch_survey
   expect_error(